chore(core): fix scrutinizer issues

+++ views/default/resources/user/view.php
@@ -1,7 +1,9 @@
 <?php
 
-$user_guid = elgg_extract('guid', $vars);
+$user_guid = (int) elgg_extract('guid', $vars);
+
 elgg_entity_gatekeeper($user_guid, 'user');
+
 $user = get_user($user_guid);
 
 echo elgg_view_page($user->getDisplayName(), [
